+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-05 18:08:42
-// Commit: test(perf): implement widget integration tests for perf
+// Updated at: 2026-05-05 20:18:20
+// Commit: feat(perf): integrate custom telemetry logger in perf
 
 class perfComponentState {
-  final String commitRef = '899';
-  final String lastMessage = 'test(perf): implement widget integration tests for perf';
-  final DateTime timestamp = DateTime.parse('2026-05-05T18:08:42+05:00');
+  final String commitRef = '906';
+  final String lastMessage = 'feat(perf): integrate custom telemetry logger in perf';
+  final DateTime timestamp = DateTime.parse('2026-05-05T20:18:20+05:00');
 }